requirements.txt
# 核心依赖
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0

# 数据增强
albumentations>=1.3.0

# 训练工具
tqdm>=4.64.0
tensorboard>=2.10.0

# 可选依赖
# scikit-learn>=1.0.0
# pytorch-lightning>=2.0.0
# wandb>=0.12.0
# hydra-core>=1.2.0
# einops>=0.5.0
# onnx>=1.12.0
